+++ src/main/java/WebProject/withpet/common/config/SecurityConfig.java
@@ -1,8 +1,8 @@
 package WebProject.withpet.common.config;
 
-import WebProject.withpet.common.auth.application.JwtTokenProvider;
-import WebProject.withpet.common.auth.filter.JwtAuthenticationFilter;
-import WebProject.withpet.common.auth.filter.JwtAuthorizationFilter;
+import WebProject.withpet.auth.application.JwtTokenProvider;
+import WebProject.withpet.auth.filter.JwtAuthenticationFilter;
+import WebProject.withpet.auth.filter.JwtAuthorizationFilter;
 import WebProject.withpet.users.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Bean;
@@ -32,35 +32,21 @@ public class SecurityConfig {
     @Bean
     public SecurityFilterChain filterChain(HttpSecurity http) throws Exception {
         AuthenticationManager authentication = authenticationManager(authenticationConfiguration);
-        JwtAuthenticationFilter authenticationFilter = new JwtAuthenticationFilter(authentication,
-            jwtTokenProvider,
-            userRepository);
+        JwtAuthenticationFilter authenticationFilter = new JwtAuthenticationFilter(authentication, jwtTokenProvider);
         authenticationFilter.setFilterProcessesUrl("/user");
 
-        http.cors().configurationSource(corsConfigurationSource())
-            .and()
-            .csrf().disable()
-            .sessionManagement()
-            .sessionCreationPolicy(SessionCreationPolicy.STATELESS)
-            .and()
+        http.cors().configurationSource(corsConfigurationSource()).and().csrf().disable().sessionManagement()
+                .sessionCreationPolicy(SessionCreationPolicy.STATELESS).and()
 
-            .formLogin().disable()
-            .httpBasic().disable()
+                .formLogin().disable().httpBasic().disable()
 
-            .addFilter(authenticationFilter)
-            .addFilter(new JwtAuthorizationFilter(authentication, jwtTokenProvider))
-            .authorizeRequests();
+                .addFilter(authenticationFilter).addFilter(new JwtAuthorizationFilter(authentication, jwtTokenProvider))
+                .authorizeRequests();
 
-        http.authorizeRequests().antMatchers("/user/**", "/mypage/**").permitAll()
-            .and()
-            .headers()
-            .addHeaderWriter(new XFrameOptionsHeaderWriter(
-                XFrameOptionsHeaderWriter.XFrameOptionsMode.SAMEORIGIN))
-            .and()
-            .logout()
-            .logoutRequestMatcher(new AntPathRequestMatcher("/user/logout"))
-            .logoutSuccessUrl("/user")
-            .invalidateHttpSession(true);
+        http.authorizeRequests().antMatchers("/user/**", "/mypage/**").permitAll().and().headers()
+                .addHeaderWriter(new XFrameOptionsHeaderWriter(XFrameOptionsHeaderWriter.XFrameOptionsMode.SAMEORIGIN))
+                .and().logout().logoutRequestMatcher(new AntPathRequestMatcher("/user/logout"))
+                .logoutSuccessUrl("/user").invalidateHttpSession(true);
 
         return http.build();
     }
@@ -71,9 +57,8 @@ public class SecurityConfig {
     }
 
     @Bean
-    public AuthenticationManager authenticationManager(
-        AuthenticationConfiguration authenticationConfiguration
-    ) throws Exception {
+    public AuthenticationManager authenticationManager(AuthenticationConfiguration authenticationConfiguration)
+            throws Exception {
         return authenticationConfiguration.getAuthenticationManager();
     }
