+++ src/main/java/WebProject/withpet/common/util/JwtUtil.java
@@ -1,13 +1,9 @@
 package WebProject.withpet.common.util;
 
-import WebProject.withpet.common.auth.application.JwtTokenProvider;
+import WebProject.withpet.auth.application.JwtTokenProvider;
 import com.auth0.jwt.JWT;
 import com.auth0.jwt.algorithms.Algorithm;
-import com.auth0.jwt.exceptions.TokenExpiredException;
-import java.io.IOException;
-import javax.servlet.http.HttpServletResponse;
 import lombok.RequiredArgsConstructor;
-import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Component;
 
 @RequiredArgsConstructor
@@ -15,6 +11,7 @@ import org.springframework.stereotype.Component;
 public class JwtUtil {
     private final JwtTokenProvider jwtTokenProvider;
 
+    // TODO : ExpiredException 발생 시 refresh token 으로 넘어가도록 수정
     public static void verify(String token, String secretKey) throws Exception {
         JWT.require(Algorithm.HMAC512(secretKey)).build().verify(token);
     }
